move main.go to be within the cmd directory structure and push all subcommands down one layer
