+++ cmd/preflight/cmd/check.go
