+++ cmd/preflight/cmd/check_operator.go
