+++ cmd/preflight/cmd/check_operator_test.go
