+++ cmd/preflight/cmd/check_test.go
