+++ cmd/preflight/cmd/runtime_assets.go
