+++ cmd/preflight/cmd/runtime_assets_test.go
