+++ cmd/preflight/cmd/support_operator_test.go
